WebM/MP4 Muxer の mux() を共通化
