+++ src/muxer/async_webm_muxer.cpp
@@ -1,18 +1,11 @@
 #include "muxer/async_webm_muxer.hpp"
 
-#include <cxxabi.h>
-#include <spdlog/spdlog.h>
-
 #include <array>
-#include <chrono>
+#include <cstdint>
 #include <filesystem>
-#include <future>
 #include <iterator>
-#include <optional>
 #include <stdexcept>
 #include <string>
-#include <system_error>
-#include <thread>
 
 #include "audio/opus.hpp"
 #include "constants.hpp"
@@ -73,101 +66,25 @@ AsyncWebMMuxer::~AsyncWebMMuxer() {
   delete m_audio_producer;
 }
 
-void AsyncWebMMuxer::addAndConsumeAudio(std::uint8_t* data,
-                                        const std::size_t data_length,
-                                        const std::uint64_t timestamp) {
-  m_context->addAudioFrame(data, data_length, timestamp);
-  delete[] data;
+void AsyncWebMMuxer::appendAudio(hisui::Frame frame) {
+  m_context->addAudioFrame(frame.data, frame.data_size, frame.timestamp);
+  delete[] frame.data;
   m_audio_producer->bufferPop();
 }
 
-void AsyncWebMMuxer::addAndConsumeVideo(std::uint8_t* data,
-                                        const std::size_t data_length,
-                                        const std::uint64_t timestamp,
-                                        const bool is_keyframe) {
-  m_context->addVideoFrame(data, data_length, timestamp, is_keyframe);
-  delete[] data;
+void AsyncWebMMuxer::appendVideo(hisui::Frame frame) {
+  m_context->addVideoFrame(frame.data, frame.data_size, frame.timestamp,
+                           frame.is_key);
+  delete[] frame.data;
   m_video_producer->bufferPop();
 }
 
 void AsyncWebMMuxer::run() {
-  auto video_future = std::async(std::launch::async, &VPXVideoProducer::produce,
-                                 m_video_producer);
-
-  auto audio_future = std::async(std::launch::async,
-                                 &OpusAudioProducer::produce, m_audio_producer);
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-  bool video_finished = false;
-
-  while (!m_audio_producer->isFinished()) {
-    const auto audio_front = m_audio_producer->bufferFront();
-    if (!audio_front.has_value()) {
-      spdlog::debug("audio queue is empty");
-      std::this_thread::sleep_for(std::chrono::milliseconds(100));
-      continue;
-    }
-    const auto audio_timestamp = audio_front.value().timestamp;
-
-    if (video_finished) {
-      addAndConsumeAudio(audio_front.value().data,
-                         audio_front.value().data_size, audio_timestamp);
-      continue;
-    }
-
-    if (m_video_producer->isFinished()) {
-      video_finished = true;
-      video_future.get();
-      spdlog::debug("video was processed");
-      addAndConsumeAudio(audio_front.value().data,
-                         audio_front.value().data_size, audio_timestamp);
-      continue;
-    }
-
-    const auto video_front = m_video_producer->bufferFront();
-    if (!video_front.has_value()) {
-      spdlog::debug("video queue is empty (1)");
-      std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-      continue;
-    }
-    const auto video_timestamp = video_front.value().timestamp;
-
-    if (video_timestamp <= audio_timestamp) {
-      addAndConsumeVideo(video_front.value().data,
-                         video_front.value().data_size, video_timestamp,
-                         video_front.value().is_key);
-      continue;
-    }
-    addAndConsumeAudio(audio_front.value().data, audio_front.value().data_size,
-                       audio_timestamp);
-  }
-
-  audio_future.get();
-  spdlog::debug("audio was processed");
-
-  if (video_finished) {
-    return;
-  }
-
-  spdlog::debug("video is processing");
-  while (!m_video_producer->isFinished()) {
-    const auto video_front = m_video_producer->bufferFront();
-    if (!video_front.has_value()) {
-      spdlog::debug("video queue is empty (2)");
-      std::this_thread::sleep_for(std::chrono::milliseconds(1000));
-      continue;
-    }
-
-    addAndConsumeVideo(video_front.value().data, video_front.value().data_size,
-                       video_front.value().timestamp,
-                       video_front.value().is_key);
-  }
-
-  video_future.get();
-  spdlog::debug("video was processed");
-}  // namespace hisui::muxer
+  mux();
+}
 
 void AsyncWebMMuxer::cleanUp() {}
+
+void AsyncWebMMuxer::muxFinalize() {}
 
 }  // namespace hisui::muxer
