+++ src/muxer/async_webm_muxer.hpp
@@ -1,12 +1,16 @@
 #pragma once
 
-#include <cstdint>
 #include <cstdio>
-#include <fstream>
 
 #include "config.hpp"
 #include "metadata.hpp"
 #include "muxer/muxer.hpp"
+
+namespace hisui {
+
+struct Frame;
+
+}
 
 namespace hisui::webm::output {
 
@@ -15,9 +19,6 @@ class Context;
 }  // namespace hisui::webm::output
 
 namespace hisui::muxer {
-
-class AudioProducer;
-class VideoProducer;
 
 class AsyncWebMMuxer : public Muxer {
  public:
@@ -29,22 +30,15 @@ class AsyncWebMMuxer : public Muxer {
   void cleanUp() override;
 
  private:
+  void muxFinalize() override;
+  void appendAudio(hisui::Frame) override;
+  void appendVideo(hisui::Frame) override;
+
   hisui::webm::output::Context* m_context;
   std::FILE* m_file;
 
   hisui::Config m_config;
   hisui::Metadata m_metadata;
-
-  VideoProducer* m_video_producer;
-  AudioProducer* m_audio_producer;
-
-  void addAndConsumeAudio(std::uint8_t*,
-                          const std::size_t,
-                          const std::uint64_t);
-  void addAndConsumeVideo(std::uint8_t*,
-                          const std::size_t,
-                          const std::uint64_t,
-                          const bool);
 };
 
 }  // namespace hisui::muxer
