+++ src/muxer/mp4_muxer.hpp
@@ -29,9 +29,6 @@ class Writer;
 
 namespace hisui::muxer {
 
-class AudioProducer;
-class VideoProducer;
-
 class MP4Muxer : public Muxer {
  public:
   ~MP4Muxer();
@@ -43,17 +40,14 @@ class MP4Muxer : public Muxer {
   shiguredo::mp4::track::SounTrack* m_soun_track;
   std::uint64_t m_chunk_interval;
 
-  VideoProducer* m_video_producer;
-  AudioProducer* m_audio_producer;
-
   std::uint64_t m_chunk_start = 0;
   std::vector<hisui::Frame> m_audio_buffer;
   std::vector<hisui::Frame> m_video_buffer;
 
-  void addAudioBuffer(hisui::Frame);
-  void addVideoBuffer(hisui::Frame);
+  void muxFinalize() override;
+  void appendAudio(hisui::Frame) override;
+  void appendVideo(hisui::Frame) override;
 
-  void mux();
   void writeTrackData();
   void initialize(const hisui::Config& t_config,
                   const hisui::Metadata& t_metadata,
